+++ .prettierrc.js
@@ -1,11 +1,11 @@
+/**
+ * To review the configuration of this file, as well as add or delete properties,
+ * consult the following link.
+ *
+ * @see https://prettier.io/docs/en/configuration.html
+ */
 module.exports = {
-  arrowParens: 'always', // Include parentheses around a sole arrow function parameter.
-  bracketSameLine: true, // Put the > of a multi-line HTML (JSX) element at the end of the last line instead of being alone on the next line.
-  bracketSpacing: true, // Print spaces between brackets in object literals.
-  semi: true, // Print semicolons at the ends of statements.
-  singleQuote: true, // Use single quotes instead of double quotes.
-  tabWidth: 2, // Specify the number of spaces per indentation-level.
-  trailingComma: 'none', // Print trailing commas wherever possible in multi-line comma-separated syntactic structures.
-  useTabs: false, // Indent lines with tabs instead of spaces.
-  quoteProps: 'as-needed' // Change when properties in objects are quoted.
+	singleQuote: true, // Use single quotes instead of double quotes.
+	trailingComma: 'none', // Print trailing commas wherever possible in multi-line comma-separated syntactic structures.
+	useTabs: true // Indent lines with tabs instead of spaces.
 };