+++ src/types.ts
@@ -1,17 +1,17 @@
 export interface SetItemLocalStorage {
-  key: string;
-  value: Record<string, any>;
+	key: string;
+	value: Record<string, any>;
 }
 
 export interface GetItemLocalStorage {
-  key: string;
-  value?: string | string[];
-  defaultValue?: any;
+	key: string;
+	value?: string | string[];
+	defaultValue?: any;
 }
 
 export interface GetItemResponseLocalStorage {
-  key?: string;
-  localStorageData?: any;
+	key?: string;
+	localStorageData?: any;
 }
 
 export type SetItemsLocalStorage = Array<SetItemLocalStorage>;